# api/swaggerdocs/swagger.yaml
basePath: /api
definitions:
  aggregations.GetHomeAggregationsResponseDto:
    description: Response for home aggregations
    properties:
      favorites:
        items:
          $ref: '#/definitions/aggregations.HServiceResponseDto'
        type: array
      featured:
        items:
          $ref: '#/definitions/aggregations.HServiceResponseDto'
        type: array
      new:
        items:
          $ref: '#/definitions/aggregations.HServiceResponseDto'
        type: array
      popular:
        items:
          $ref: '#/definitions/aggregations.HServiceResponseDto'
        type: array
    type: object
  aggregations.HServiceResponseDto:
    description: Basic service information with user information
    properties:
      category:
        example: 1
        type: integer
      createdAt:
        example: "2024-08-26T10:24:13.508676+03:00"
        type: string
      deliveryTime:
        example: 1
        type: integer
      deliveryTimespan:
        example: HOURLY
        type: string
      description:
        example: Example service description
        type: string
      id:
        example: "7234882566245847040"
        type: string
      isOnline:
        example: true
        type: boolean
      location:
        example: Example Location
        type: string
      media:
        additionalProperties: {}
        type: object
      price:
        example: 10
        type: number
      priceTimespan:
        example: HOURLY
        type: string
      priceUnit:
        example: USD
        type: string
      slug:
        example: example-service
        type: string
      title:
        example: Example Service
        type: string
      totalPoints:
        example: 50
        type: integer
      totalVotes:
        example: 10
        type: integer
      updatedAt:
        example: "2024-08-26T10:24:13.508676+03:00"
        type: string
      url:
        example: https://example.com
        type: string
      user:
        $ref: '#/definitions/aggregations.UserResponseDto'
      userId:
        example: "528696135489945615"
        type: string
    type: object
  aggregations.UserResponseDto:
    description: Basic user information
    properties:
      createdAt:
        example: "2024-08-26T10:24:13.508676+03:00"
        type: string
      fullName:
        example: John Doe
        type: string
      gender:
        example: male
        type: string
      id:
        example: "528696135489945615"
        type: string
      profileImage:
        example: https://example.com/image.jpg
        type: string
      username:
        example: johndoe
        type: string
    type: object
  auth.GetMeResponseDto:
    properties:
      createdAt:
        example: "2024-08-26T10:24:13.508676+03:00"
        format: date-time
        type: string
      email:
        example: johndoe@example.com
        type: string
      fullName:
        example: John Doe
        type: string
      gender:
        example: male
        type: string
      googleId:
        example: "10887502189381205719451"
        type: string
      id:
        example: "528696135489945615"
        type: string
      isActive:
        example: true
        type: boolean
      isEmailVerified:
        example: true
        type: boolean
      lastLogin:
        example: "2024-08-26T10:24:13.508676+03:00"
        type: string
      profileImage:
        example: https://example.com/image.jpg
        type: string
      role:
        example: user
        type: string
      updatedAt:
        example: "2024-08-26T10:24:13.508676+03:00"
        format: date-time
        type: string
      username:
        example: johndoe
        type: string
    type: object
  auth.LoginRequestDto:
    description: Login request dto
    properties:
      email:
        example: johndoe@example.com
        type: string
      password:
        example: password123
        format: password
        maxLength: 128
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequestDto:
    properties:
      email:
        example: johndoe@example.com
        maxLength: 128
        minLength: 3
        type: string
      fullName:
        example: John Doe
        maxLength: 128
        minLength: 3
        type: string
      password:
        example: password123
        format: password
        maxLength: 128
        minLength: 6
        type: string
      username:
        example: johndoe
        maxLength: 32
        minLength: 4
        type: string
    required:
    - email
    - fullName
    - password
    - username
    type: object
  auth.ResetPasswordRequestDto:
    properties:
      code:
        example: "123456"
        type: string
      email:
        example: johndoe@example.com
        type: string
      newPassword:
        example: password123
        format: password
        maxLength: 128
        minLength: 6
        type: string
    required:
    - code
    - email
    - newPassword
    type: object
  auth.SendForgotPasswordEmailRequestDto:
    properties:
      email:
        example: johndoe@example.com
        type: string
    required:
    - email
    type: object
  auth.SendVerificationEmailRequestDto:
    properties:
      email:
        example: johndoe@example.com
        type: string
    required:
    - email
    type: object
  echo.HTTPError:
    properties:
      message: {}
    type: object
  health.GetHealthResponseDto:
    description: Generic response dto
    properties:
      message:
        type: string
    type: object
host: localhost:5000
info:
  contact:
    email: gethorizonapp@gmail.com
    name: Mert Turkmenoglu
    url: https://mertturkmenoglu.com
  description: Horizon backend services
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: http://localhost:3000/terms
  title: Horizon API
  version: "1.0"
paths:
  /aggregations/home:
    get:
      consumes:
      - application/json
      description: An endpoint to fetch multiple homepage queries
      produces:
      - application/json
      responses:
        "200":
          description: Successful request
          schema:
            $ref: '#/definitions/aggregations.GetHomeAggregationsResponseDto'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Fetch multiple homepage entities in a single request
      tags:
      - Aggregations
  /auth/credentials/login:
    post:
      consumes:
      - application/json
      description: Logs in the user with email and password
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequestDto'
      responses:
        "200":
          description: OK
        "400":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Login with email and password
      tags:
      - Auth
  /auth/credentials/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with email and password
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequestDto'
      responses:
        "201":
          description: Created
        "400":
          description: Invalid email or username
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Register with email and password
      tags:
      - Auth
  /auth/forgot-password/reset:
    post:
      consumes:
      - application/json
      description: Resets the password of the user
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordRequestDto'
      responses:
        "200":
          description: OK
        "400":
          description: Invalid email or code
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Reset password
      tags:
      - Auth
  /auth/forgot-password/send:
    post:
      consumes:
      - application/json
      description: Sends a forgot password email to the user
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.SendForgotPasswordEmailRequestDto'
      responses:
        "200":
          description: OK
        "400":
          description: Invalid email
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Send forgot password email
      tags:
      - Auth
  /auth/google:
    get:
      consumes:
      - application/json
      description: Login with Google OAuth2
      responses:
        "307":
          description: Temporary Redirect
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Login with Google OAuth2
      tags:
      - Auth
  /auth/google/callback:
    get:
      consumes:
      - application/json
      description: Google OAuth2 callback
      responses:
        "307":
          description: Temporary Redirect
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Google OAuth2 callback
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out the current user
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Logs out the current user
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Gets the currently authenticated user or returns an error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.GetMeResponseDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Gets the current user
      tags:
      - Auth
  /auth/verify-email/send:
    post:
      consumes:
      - application/json
      description: Sends a verification email to the user
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.SendVerificationEmailRequestDto'
      responses:
        "200":
          description: OK
        "400":
          description: Invalid email or email already verified
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Send verification email
      tags:
      - Auth
  /auth/verify-email/verify:
    get:
      consumes:
      - application/json
      description: Verifies the email of the user
      parameters:
      - description: Verification code
        in: query
        name: code
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Invalid or expired verification code
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Verify email
      tags:
      - Auth
  /health/:
    get:
      consumes:
      - application/json
      description: An endpoint to be used by load balancers to check the health of
        the service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/health.GetHealthResponseDto'
      summary: Returns ok message
      tags:
      - health
swagger: "2.0"